requirements.txt
Django==4.0.5
djangorestframework==3.13.1
psycopg2==2.9.3
# psycopg2-binary-2.9.3
Pillow==9.1.1
django-elasticsearch-dsl==7.2.2